No longer have to call Build to create the instance
Now uses an implicit conversion operator. Note that interfaces can't have these, so we don't deal in IObjectBuilders anymore; instead, it's now a concrete ParameterComposer. Ditched IObjectBuilder entirely.
