+++ src/NanoBuilder/NanoBuilder/NanoBuilder.csproj
@@ -49,7 +49,7 @@
   <ItemGroup>
     <Compile Include="AmbiguousConstructorException.cs" />
     <Compile Include="EnumerableExtensions.cs" />
-    <Compile Include="IObjectBuilder.cs" />
+    <Compile Include="ParameterComposer.cs" />
     <Compile Include="ITypeInspector.cs" />
     <Compile Include="ITypeMapper.cs" />
     <Compile Include="MoqMapper.cs" />
